requirements.txt
requests==2.32.3
PyYAML==6.0.2
pysigma==0.11.16
pysigma-backend-cortexxdr==0.4.2
git+https://github.com/gocortexio/sigma2xsiam.git
